Refactored box component to separate styling from business logic.
